replace redundant index.js/<ModuleName>.js pattern for just <ModuleFolder>/index.js
