--- app/src/components/routes/navigation/Navigation.js
@@ -1,9 +1,0 @@
-import React from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import DrawerNavigator from "./drawer";
-
-export default () => (
-  <NavigationContainer>
-    <DrawerNavigator />
-  </NavigationContainer>
-);
