+++ app/src/components/routes/navigation/index.js
@@ -1,3 +1,9 @@
-import Navigation from "./Navigation";
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import DrawerNavigator from "./drawer";
 
-export default Navigation;
+export default () => (
+  <NavigationContainer>
+    <DrawerNavigator />
+  </NavigationContainer>
+);
